Refactor name input
